requirements.txt
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.26
chromadb==1.0.9
langchain-core==0.3.66
langchain-experimental==0.3.4
langchain-ollama==0.3.3
langchain-openai==0.3.27
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
fastapi
uvicorn==0.34.0
sqlalchemy
asyncpg==0.29.0
alembic==1.13.1
greenlet